#2281 Move everything about `date` conditions to the `date` package
