+++ src/main/java/com/codeborne/selenide/Condition.java
@@ -5,15 +5,6 @@ import com.codeborne.selenide.conditions.Attribute;
 import com.codeborne.selenide.conditions.AttributeWithValue;
 import com.codeborne.selenide.conditions.CaseSensitiveText;
 import com.codeborne.selenide.conditions.Checked;
-import com.codeborne.selenide.conditions.InnerText;
-import com.codeborne.selenide.conditions.LocalDateBetweenCondition;
-import com.codeborne.selenide.conditions.LocalDateCondition;
-import com.codeborne.selenide.conditions.LocalDateFormatCondition;
-import com.codeborne.selenide.conditions.LocalDateTimeBetweenCondition;
-import com.codeborne.selenide.conditions.LocalDateTimeCondition;
-import com.codeborne.selenide.conditions.LocalDateTimeFormatCondition;
-import com.codeborne.selenide.conditions.PartialText;
-import com.codeborne.selenide.conditions.PartialTextCaseSensitive;
 import com.codeborne.selenide.conditions.CssClass;
 import com.codeborne.selenide.conditions.CssValue;
 import com.codeborne.selenide.conditions.CustomMatch;
@@ -29,6 +20,7 @@ import com.codeborne.selenide.conditions.ExplainedCondition;
 import com.codeborne.selenide.conditions.Focused;
 import com.codeborne.selenide.conditions.Hidden;
 import com.codeborne.selenide.conditions.Href;
+import com.codeborne.selenide.conditions.InnerText;
 import com.codeborne.selenide.conditions.Interactable;
 import com.codeborne.selenide.conditions.IsImageLoaded;
 import com.codeborne.selenide.conditions.MatchAttributeWithValue;
@@ -38,6 +30,8 @@ import com.codeborne.selenide.conditions.Not;
 import com.codeborne.selenide.conditions.Or;
 import com.codeborne.selenide.conditions.OwnText;
 import com.codeborne.selenide.conditions.OwnTextCaseSensitive;
+import com.codeborne.selenide.conditions.PartialText;
+import com.codeborne.selenide.conditions.PartialTextCaseSensitive;
 import com.codeborne.selenide.conditions.PartialValue;
 import com.codeborne.selenide.conditions.PseudoElementPropertyWithValue;
 import com.codeborne.selenide.conditions.Readonly;
@@ -47,6 +41,12 @@ import com.codeborne.selenide.conditions.TagName;
 import com.codeborne.selenide.conditions.Text;
 import com.codeborne.selenide.conditions.Value;
 import com.codeborne.selenide.conditions.Visible;
+import com.codeborne.selenide.conditions.date.LocalDateBetweenCondition;
+import com.codeborne.selenide.conditions.date.LocalDateCondition;
+import com.codeborne.selenide.conditions.date.LocalDateFormatCondition;
+import com.codeborne.selenide.conditions.date.LocalDateTimeBetweenCondition;
+import com.codeborne.selenide.conditions.date.LocalDateTimeCondition;
+import com.codeborne.selenide.conditions.date.LocalDateTimeFormatCondition;
 import org.openqa.selenium.WebElement;
 
 import javax.annotation.CheckReturnValue;
