+++ src/test/java/com/codeborne/selenide/conditions/date/LocalDateBetweenConditionTest.java
@@ -1,11 +1,11 @@
-package com.codeborne.selenide.conditions;
+package com.codeborne.selenide.conditions.date;
 
 import com.codeborne.selenide.CheckResult;
 import com.codeborne.selenide.Driver;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.WebElement;
 
-import java.time.LocalDateTime;
+import java.time.LocalDate;
 
 import static com.codeborne.selenide.CheckResult.Verdict.ACCEPT;
 import static com.codeborne.selenide.CheckResult.Verdict.REJECT;
@@ -16,17 +16,17 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-final class LocalDateTimeBetweenConditionTest {
+final class LocalDateBetweenConditionTest {
   private final Driver driver = mock();
   private final WebElement element = mock();
-  private final LocalDateTimeFormatCondition formatCondition = mock();
-  private final LocalDateTime start = LocalDateTime.of(2022, 10, 9, 10, 10, 10);
-  private final LocalDateTime end = LocalDateTime.of(2022, 10, 12, 10, 10, 10);
-  private final LocalDateTimeBetweenCondition condition = new LocalDateTimeBetweenCondition(start, end, formatCondition);
+  private final LocalDateFormatCondition formatCondition = mock();
+  private final LocalDate start = LocalDate.of(2022, 10, 9);
+  private final LocalDate end = LocalDate.of(2022, 10, 12);
+  private final LocalDateBetweenCondition condition = new LocalDateBetweenCondition(start, end, formatCondition);
 
   @Test
   void correctDateValueWithCorrectFormat() {
-    when(formatCondition.check(any(), any())).thenReturn(new CheckResult(ACCEPT, LocalDateTime.of(2022, 10, 11, 10, 10, 10)));
+    when(formatCondition.check(any(), any())).thenReturn(new CheckResult(ACCEPT, LocalDate.of(2022, 10, 11)));
     when(formatCondition.format(any())).thenReturn("formatted date");
 
     CheckResult check = condition.check(driver, element);
@@ -35,7 +35,7 @@ final class LocalDateTimeBetweenConditionTest {
     assertThat(check.actualValue()).isEqualTo("formatted date");
 
     verify(formatCondition).check(driver, element);
-    verify(formatCondition).format(LocalDateTime.of(2022, 10, 11, 10, 10, 10));
+    verify(formatCondition).format(LocalDate.of(2022, 10, 11));
   }
 
   @Test
@@ -50,7 +50,7 @@ final class LocalDateTimeBetweenConditionTest {
 
   @Test
   void incorrectDateValue() {
-    when(formatCondition.check(any(), any())).thenReturn(new CheckResult(ACCEPT, LocalDateTime.of(2021, 10, 12, 10, 10, 10)));
+    when(formatCondition.check(any(), any())).thenReturn(new CheckResult(ACCEPT, LocalDate.of(2021, 10, 12)));
     when(formatCondition.format(any())).thenReturn("formatted date");
 
     CheckResult check = condition.check(driver, element);
@@ -59,32 +59,31 @@ final class LocalDateTimeBetweenConditionTest {
     assertThat(check.actualValue()).isEqualTo("formatted date");
 
     verify(formatCondition).check(driver, element);
-    verify(formatCondition).format(LocalDateTime.of(2021, 10, 12, 10, 10, 10));
+    verify(formatCondition).format(LocalDate.of(2021, 10, 12));
   }
 
   @Test
   void stringRepresentationOfCondition() {
-    var condition = new LocalDateTimeBetweenCondition(start, end, "yyyy/MM/dd HH:mm:ss");
+    var condition = new LocalDateBetweenCondition(start, end, "yyyy/MM/dd");
 
     assertThat(condition.toString())
-      .isEqualTo("datetime value between [\"2022/10/09 10:10:10\", \"2022/10/12 10:10:10\"] " +
-        "(with datetime value format: \"yyyy/MM/dd HH:mm:ss\")");
+      .isEqualTo("date value between [\"2022/10/09\", \"2022/10/12\"] (with date value format: \"yyyy/MM/dd\")");
   }
 
   @Test
   void throwExceptionIfStartDateIsEqualToEndDate() {
-    LocalDateTime date = LocalDateTime.of(2022, 10, 11, 10, 10, 10);
-    assertThatThrownBy(() -> new LocalDateTimeBetweenCondition(date, date, formatCondition))
+    LocalDate date = LocalDate.of(2022, 10, 11);
+    assertThatThrownBy(() -> new LocalDateBetweenCondition(date, date, formatCondition))
       .isInstanceOf(IllegalArgumentException.class)
-      .hasMessage("startDateTime must be before endDateTime");
+      .hasMessage("startDate must be before endDate");
   }
 
   @Test
   void throwExceptionIfStartDateAfterEndDate() {
-    LocalDateTime startDate = LocalDateTime.of(2022, 10, 11, 10, 10, 10);
-    LocalDateTime endDate = LocalDateTime.of(2021, 10, 11, 10, 10, 10);
-    assertThatThrownBy(() -> new LocalDateTimeBetweenCondition(startDate, endDate, formatCondition))
+    LocalDate startDate = LocalDate.of(2022, 10, 11);
+    LocalDate endDate = LocalDate.of(2021, 10, 11);
+    assertThatThrownBy(() -> new LocalDateBetweenCondition(startDate, endDate, formatCondition))
       .isInstanceOf(IllegalArgumentException.class)
-      .hasMessage("startDateTime must be before endDateTime");
+      .hasMessage("startDate must be before endDate");
   }
 }
