+++ src/test/java/com/codeborne/selenide/conditions/date/LocalDateConditionTest.java
@@ -1,11 +1,11 @@
-package com.codeborne.selenide.conditions;
+package com.codeborne.selenide.conditions.date;
 
 import com.codeborne.selenide.CheckResult;
 import com.codeborne.selenide.Driver;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.WebElement;
 
-import java.time.LocalDateTime;
+import java.time.LocalDate;
 
 import static com.codeborne.selenide.CheckResult.Verdict.ACCEPT;
 import static com.codeborne.selenide.CheckResult.Verdict.REJECT;
@@ -15,15 +15,15 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-final class LocalDateTimeConditionTest {
+final class LocalDateConditionTest {
   private final Driver driver = mock();
   private final WebElement element = mock();
-  private final LocalDateTimeFormatCondition formatCondition = mock();
-  private final LocalDateTime expectedDateTime = LocalDateTime.of(2022, 10, 11, 12, 13, 14);
-  private final LocalDateTimeCondition condition = new LocalDateTimeCondition(expectedDateTime, formatCondition);
+  private final LocalDateFormatCondition formatCondition = mock();
+  private final LocalDateCondition condition = new LocalDateCondition(LocalDate.of(2022, 10, 11), formatCondition);
+
   @Test
-  void correctDateTimeValueWithCorrectFormat() {
-    when(formatCondition.check(any(), any())).thenReturn(new CheckResult(ACCEPT, expectedDateTime));
+  void correctDateValueWithCorrectFormat() {
+    when(formatCondition.check(any(), any())).thenReturn(new CheckResult(ACCEPT, LocalDate.of(2022, 10, 11)));
     when(formatCondition.format(any())).thenReturn("formatted date");
 
     CheckResult check = condition.check(driver, element);
@@ -32,7 +32,7 @@ final class LocalDateTimeConditionTest {
     assertThat(check.actualValue()).isEqualTo("formatted date");
 
     verify(formatCondition).check(driver, element);
-    verify(formatCondition).format(expectedDateTime);
+    verify(formatCondition).format(LocalDate.of(2022, 10, 11));
   }
 
   @Test
@@ -46,8 +46,8 @@ final class LocalDateTimeConditionTest {
   }
 
   @Test
-  void incorrectDateTimeValue() {
-    when(formatCondition.check(any(), any())).thenReturn(new CheckResult(ACCEPT, expectedDateTime.minusDays(1)));
+  void incorrectDateValue() {
+    when(formatCondition.check(any(), any())).thenReturn(new CheckResult(ACCEPT, LocalDate.of(2023, 10, 11)));
     when(formatCondition.format(any())).thenReturn("formatted date");
 
     CheckResult check = condition.check(driver, element);
@@ -56,14 +56,13 @@ final class LocalDateTimeConditionTest {
     assertThat(check.actualValue()).isEqualTo("formatted date");
 
     verify(formatCondition).check(driver, element);
-    verify(formatCondition).format(expectedDateTime.minusDays(1));
+    verify(formatCondition).format(LocalDate.of(2023, 10, 11));
   }
 
   @Test
   void stringRepresentationOfCondition() {
-    var condition = new LocalDateTimeCondition(expectedDateTime, "yyyy/MM/dd HH:mm:ss");
+    var condition = new LocalDateCondition(LocalDate.of(2022, 10, 11), "yyyy/MM/dd");
 
-    assertThat(condition.toString())
-      .isEqualTo("datetime value: \"2022/10/11 12:13:14\" (with datetime value format: \"yyyy/MM/dd HH:mm:ss\")");
+    assertThat(condition.toString()).isEqualTo("date value: \"2022/10/11\" (with date value format: \"yyyy/MM/dd\")");
   }
 }
