+++ src/test/java/com/codeborne/selenide/conditions/date/LocalDateTimeBetweenConditionTest.java
@@ -1,11 +1,11 @@
-package com.codeborne.selenide.conditions;
+package com.codeborne.selenide.conditions.date;
 
 import com.codeborne.selenide.CheckResult;
 import com.codeborne.selenide.Driver;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.WebElement;
 
-import java.time.LocalDate;
+import java.time.LocalDateTime;
 
 import static com.codeborne.selenide.CheckResult.Verdict.ACCEPT;
 import static com.codeborne.selenide.CheckResult.Verdict.REJECT;
@@ -16,17 +16,17 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-final class LocalDateBetweenConditionTest {
+final class LocalDateTimeBetweenConditionTest {
   private final Driver driver = mock();
   private final WebElement element = mock();
-  private final LocalDateFormatCondition formatCondition = mock();
-  private final LocalDate start = LocalDate.of(2022, 10, 9);
-  private final LocalDate end = LocalDate.of(2022, 10, 12);
-  private final LocalDateBetweenCondition condition = new LocalDateBetweenCondition(start, end, formatCondition);
+  private final LocalDateTimeFormatCondition formatCondition = mock();
+  private final LocalDateTime start = LocalDateTime.of(2022, 10, 9, 10, 10, 10);
+  private final LocalDateTime end = LocalDateTime.of(2022, 10, 12, 10, 10, 10);
+  private final LocalDateTimeBetweenCondition condition = new LocalDateTimeBetweenCondition(start, end, formatCondition);
 
   @Test
   void correctDateValueWithCorrectFormat() {
-    when(formatCondition.check(any(), any())).thenReturn(new CheckResult(ACCEPT, LocalDate.of(2022, 10, 11)));
+    when(formatCondition.check(any(), any())).thenReturn(new CheckResult(ACCEPT, LocalDateTime.of(2022, 10, 11, 10, 10, 10)));
     when(formatCondition.format(any())).thenReturn("formatted date");
 
     CheckResult check = condition.check(driver, element);
@@ -35,7 +35,7 @@ final class LocalDateBetweenConditionTest {
     assertThat(check.actualValue()).isEqualTo("formatted date");
 
     verify(formatCondition).check(driver, element);
-    verify(formatCondition).format(LocalDate.of(2022, 10, 11));
+    verify(formatCondition).format(LocalDateTime.of(2022, 10, 11, 10, 10, 10));
   }
 
   @Test
@@ -50,7 +50,7 @@ final class LocalDateBetweenConditionTest {
 
   @Test
   void incorrectDateValue() {
-    when(formatCondition.check(any(), any())).thenReturn(new CheckResult(ACCEPT, LocalDate.of(2021, 10, 12)));
+    when(formatCondition.check(any(), any())).thenReturn(new CheckResult(ACCEPT, LocalDateTime.of(2021, 10, 12, 10, 10, 10)));
     when(formatCondition.format(any())).thenReturn("formatted date");
 
     CheckResult check = condition.check(driver, element);
@@ -59,31 +59,32 @@ final class LocalDateBetweenConditionTest {
     assertThat(check.actualValue()).isEqualTo("formatted date");
 
     verify(formatCondition).check(driver, element);
-    verify(formatCondition).format(LocalDate.of(2021, 10, 12));
+    verify(formatCondition).format(LocalDateTime.of(2021, 10, 12, 10, 10, 10));
   }
 
   @Test
   void stringRepresentationOfCondition() {
-    var condition = new LocalDateBetweenCondition(start, end, "yyyy/MM/dd");
+    var condition = new LocalDateTimeBetweenCondition(start, end, "yyyy/MM/dd HH:mm:ss");
 
     assertThat(condition.toString())
-      .isEqualTo("date value between [\"2022/10/09\", \"2022/10/12\"] (with date value format: \"yyyy/MM/dd\")");
+      .isEqualTo("datetime value between [\"2022/10/09 10:10:10\", \"2022/10/12 10:10:10\"] " +
+        "(with datetime value format: \"yyyy/MM/dd HH:mm:ss\")");
   }
 
   @Test
   void throwExceptionIfStartDateIsEqualToEndDate() {
-    LocalDate date = LocalDate.of(2022, 10, 11);
-    assertThatThrownBy(() -> new LocalDateBetweenCondition(date, date, formatCondition))
+    LocalDateTime date = LocalDateTime.of(2022, 10, 11, 10, 10, 10);
+    assertThatThrownBy(() -> new LocalDateTimeBetweenCondition(date, date, formatCondition))
       .isInstanceOf(IllegalArgumentException.class)
-      .hasMessage("startDate must be before endDate");
+      .hasMessage("startDateTime must be before endDateTime");
   }
 
   @Test
   void throwExceptionIfStartDateAfterEndDate() {
-    LocalDate startDate = LocalDate.of(2022, 10, 11);
-    LocalDate endDate = LocalDate.of(2021, 10, 11);
-    assertThatThrownBy(() -> new LocalDateBetweenCondition(startDate, endDate, formatCondition))
+    LocalDateTime startDate = LocalDateTime.of(2022, 10, 11, 10, 10, 10);
+    LocalDateTime endDate = LocalDateTime.of(2021, 10, 11, 10, 10, 10);
+    assertThatThrownBy(() -> new LocalDateTimeBetweenCondition(startDate, endDate, formatCondition))
       .isInstanceOf(IllegalArgumentException.class)
-      .hasMessage("startDate must be before endDate");
+      .hasMessage("startDateTime must be before endDateTime");
   }
 }
